LogsMetaRow: extract download function to logs utils
